한수연 > 장바구니 완성

+++ CCcaseShoppingMall/src/member/controller/MyCartDeleteAction.java
@@ -7,6 +7,7 @@ import org.json.JSONObject;
 
 import common.controller.AbstractController;
 import member.model.CartDAO;
+import member.model.CartVO;
 import member.model.InterCartDAO;
 
 public class MyCartDeleteAction extends AbstractController {
@@ -30,19 +31,36 @@ public class MyCartDeleteAction extends AbstractController {
 		}
 		else { // post방식으로 들어온 경우
 			
-			int cartno= Integer.parseInt(request.getParameter("cartno"));
+			String scartno= request.getParameter("cartno");
+			String fk_productid= request.getParameter("productid");
+			String fk_pnum= request.getParameter("pnum"); // 색상선택 안한 경우 "-"
+			String cinputcnt= request.getParameter("pcnt");
+			String fk_userid= request.getParameter("userid");
 			
-			InterCartDAO ctdao= new CartDAO();
-			int n= ctdao.deleteOneRow(cartno);
 			
-			JSONObject jsonObject= new JSONObject();
-			jsonObject.put("n", n);
-			
-			String json= jsonObject.toString();
-			request.setAttribute("json", json);
-			
-			super.setRedirect(false);
-			super.setViewPage("/WEB-INF/jsonview.jsp");
+			if(scartno!="null") {
+				
+				int cartno= Integer.parseInt(scartno);
+				InterCartDAO ctdao= new CartDAO();
+				int n= ctdao.deleteOneRow(cartno);
+				
+				CartVO ctvo= new CartVO();
+				ctvo.setCinputcnt(Integer.parseInt(cinputcnt));
+				ctvo.setFk_pnum(fk_pnum);
+				ctvo.setFk_productid(fk_productid);
+				ctvo.setFk_userid(fk_userid);
+				
+				n= ctdao.insertUpdateWishList(ctvo);
+				
+				JSONObject jsonObject= new JSONObject();
+				jsonObject.put("n", n);
+				
+				String json= jsonObject.toString();
+				request.setAttribute("json", json);
+				
+				super.setRedirect(false);
+				super.setViewPage("/WEB-INF/jsonview.jsp");
+			}
 		}
 	}
